+++ con-rs/Cargo.toml
@@ -8,3 +8,4 @@ serde = { version = "1", features = ["derive"] }
 rand = "0.8"
 blake2b_simd = "1"
 tracing = "0.1"
+minicbor = { version = "0.25", features = ["std"] }
