+++ con-rs/src/lib.rs
@@ -7,9 +7,12 @@
 //! aggregation / quorum detection.
 
 pub mod aggregation;
+pub mod chain_tree;
 pub mod config;
 pub mod elections;
 pub mod pipeline;
+pub mod types;
 pub mod wfa;
 
 pub use config::{CommitteeSelection, StakeEntry};
+pub use types::Point;
