+++ net-rs/net-core/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+con-rs = { path = "../../con-rs" }
 tokio = { version = "1", features = ["rt", "net", "io-util", "time", "sync", "macros"] }
 minicbor = { version = "0.25", features = ["std"] }
 blake2b_simd = "1"
