+++ net-rs/net-core/src/types/mod.rs
@@ -16,6 +16,8 @@ use minicbor::decode::Error as DecodeError;
 use minicbor::encode::Error as EncodeError;
 use minicbor::{Decode, Decoder, Encode, Encoder};
 
+pub use con_rs::Point;
+
 /// Maximum number of points in a FindIntersect message.
 pub const MAX_POINTS: usize = 2048;
 
@@ -24,98 +26,6 @@ pub const MAX_HEADER_SIZE: usize = 65_535;
 
 /// Maximum block body size (matches BlockFetch StStreaming size limit).
 pub const MAX_BLOCK_SIZE: usize = 2_500_000;
-
-// --- Point ---
-
-/// A point on the chain: either the genesis (origin) or a specific slot+hash.
-///
-/// Wire format:
-///   origin   = []                    (empty array)
-///   specific = [slotNo, headerHash]  (2-element array)
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub enum Point {
-    Origin,
-    Specific { slot: u64, hash: [u8; 32] },
-}
-
-impl fmt::Display for Point {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Point::Origin => write!(f, "origin"),
-            Point::Specific { slot, hash } => {
-                write!(f, "{}/{}", slot, hex_prefix(hash))
-            }
-        }
-    }
-}
-
-impl minicbor::Encode<()> for Point {
-    fn encode<W: minicbor::encode::Write>(
-        &self,
-        e: &mut Encoder<W>,
-        _ctx: &mut (),
-    ) -> Result<(), EncodeError<W::Error>> {
-        match self {
-            Point::Origin => {
-                e.array(0)?;
-            }
-            Point::Specific { slot, hash } => {
-                e.array(2)?;
-                e.u64(*slot)?;
-                e.bytes(hash)?;
-            }
-        }
-        Ok(())
-    }
-}
-
-impl<'a> minicbor::Decode<'a, ()> for Point {
-    fn decode(d: &mut Decoder<'a>, _ctx: &mut ()) -> Result<Self, DecodeError> {
-        let len = d.array()?;
-        match len {
-            Some(0) => Ok(Point::Origin),
-            None => {
-                // Indefinite-length array — check if immediately closed (origin)
-                // or has elements (specific).
-                if d.datatype()? == minicbor::data::Type::Break {
-                    d.skip()?; // consume the break
-                    return Ok(Point::Origin);
-                }
-                let slot = d.u64()?;
-                let hash_bytes = d.bytes()?;
-                if hash_bytes.len() != 32 {
-                    return Err(DecodeError::message(format!(
-                        "point hash must be 32 bytes, got {}",
-                        hash_bytes.len()
-                    )));
-                }
-                let mut hash = [0u8; 32];
-                hash.copy_from_slice(hash_bytes);
-                // Consume the break marker.
-                if d.datatype()? == minicbor::data::Type::Break {
-                    d.skip()?;
-                }
-                Ok(Point::Specific { slot, hash })
-            }
-            Some(2) => {
-                let slot = d.u64()?;
-                let hash_bytes = d.bytes()?;
-                if hash_bytes.len() != 32 {
-                    return Err(DecodeError::message(format!(
-                        "point hash must be 32 bytes, got {}",
-                        hash_bytes.len()
-                    )));
-                }
-                let mut hash = [0u8; 32];
-                hash.copy_from_slice(hash_bytes);
-                Ok(Point::Specific { slot, hash })
-            }
-            Some(other) => Err(DecodeError::message(format!(
-                "expected point array of length 0 or 2, got {other}"
-            ))),
-        }
-    }
-}
 
 // --- Tip ---
 
@@ -157,14 +67,6 @@ impl<'a> minicbor::Decode<'a, ()> for Tip {
 }
 
 // --- Helpers ---
-
-/// Hex-encode the first 8 bytes of a hash for display.
-fn hex_prefix(hash: &[u8; 32]) -> String {
-    hash.iter()
-        .take(8)
-        .map(|b| format!("{b:02x}"))
-        .collect::<String>()
-}
 
 /// Decode an array of points, handling both definite and indefinite length.
 /// Enforces MAX_POINTS to prevent unbounded allocation.
