+++ net-rs/net-node/src/main.rs
@@ -1,4 +1,3 @@
-mod chain_tree;
 mod clock;
 mod config;
 mod consensus;
